fix: refactored code AM 306
