+++ src/test/resources/exception-mapper-data/malformedInput.json
@@ -1,0 +1,17 @@
+{
+  "userId": "1234",
+  "userRoles": [
+    "caseworker"
+  ],
+  "resource": {
+    "id": "1234",
+    "definition": {
+      "serviceName": "cmc",
+      "resourceType": "case",
+      "resourceName": "claim"
+    },
+    "data": {
+      "json": "resource"
+    }
+  },,,
+}
