+++ src/test/resources/input-data/createResourceAccess.json
@@ -1,0 +1,14 @@
+{
+  "resourceId": "1234",
+  "resourceDefinition": {
+    "serviceName": "cmc",
+    "resourceType": "case",
+    "resourceName": "claim"
+  },
+  "accessorIds": ["12345"],
+  "accessorType": "USER",
+  "attributePermissions": {
+    "": ["CREATE", "READ", "UPDATE"]
+  },
+  "relationship": "caseworker"
+}
